pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "infra"
version = "0.1.0"
description = "Infrastructure automation toolkit for rapid project deployment"
readme = "README.md"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
repository = "https://github.com/yourusername/infra"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
packages = [{include = "infra"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
click = ">=8.1.0"
python-dotenv = ">=1.0.0"
requests = ">=2.28.0"
PyGithub = ">=2.0.0"
pyyaml = ">=6.0"
paramiko = ">=3.0.0"   # For SSH operations
yandexcloud = ">=0.212.0"  # For Yandex Cloud
jinja2 = ">=3.1.0"     # For templating
gitpython = ">=3.1.30"  # For Git operations

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"
mypy = ">=1.0.0"

[tool.poetry.group.gui.dependencies]
PyQt6 = ">=6.4.0"      # For macOS GUI
py2app = ">=0.28.0"    # For building macOS app bundle

[tool.poetry.group.ai.dependencies]
openai = ">=1.0.0"     # For AI-assisted automation

[tool.poetry.scripts]
infra = "infra.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 